Relocate clone package
