+++ clone/clone.go
