+++ clone/clone_test.go
