+++ transform/resize.go
@@ -4,7 +4,7 @@ import (
 	"image"
 	"math"
 
-	"github.com/anthonynsimon/bild/image/clone"
+	"github.com/anthonynsimon/bild/clone"
 	"github.com/anthonynsimon/bild/math/f64"
 	"github.com/anthonynsimon/bild/parallel"
 )
